# backend/openapi.yaml
openapi: 3.0.1
info:
  title: E-COM PORTFOLIO CODECADDEMY PROJECT
  version: 1.0.0
  description: >-
    Part 1 of the the main portfolio project on codecademys fullstack engineer
    course
paths:
...
  /orders/order/{id}/summary:
    summary: 'Gets an orders from the db and displays an overall summary '
    description: 'This path is used to retrieve an order from the db '
    get:
      summary: Gets an order from the db to display an order summary
      description: Retries an orer from the db
      operationId: get_order_summary
      parameters:
        - name: id
          in: path
          description: the id of an order
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Request successful. The server has responded as required.
          content:
            application/json:
              examples:
                get_order_summary:
                  value: |-
                    {
                        "data": {
                            "user_id": 1,
                            "orderTotal": "55.00",
                            "items": [
                                {
                                    "itemId": 2,
                                    "name": "black bomber jacket",
                                    "price": "17.00",
                                    "quantity": 1,
                                    "size": "L",
                                    "price_X_quantity": "17.00"
                                },
                                {
                                    "itemId": 1,
                                    "name": "nike running shoes",
                                    "price": "38.00",
                                    "quantity": 1,
                                    "size": "8.5",
                                    "price_X_quantity": "38.00"
                                }
                            ]
                        }
                    }
      tags:
        - orders
...